chore(queue): moves queue logic files
