+++ modules/queue/logic/messageBrokerLogic.js
@@ -1,19 +1,21 @@
-const { MESSAGE_QUEUES } = require('../models/enums/queues');
-const queue = require('../utils/queue');
+const { MESSAGE_QUEUES } = require('../../../models/enums/queues');
+const queue = require('./queueLogic');
 
-const broker = {
+class MessageBroker {
   async init() {
     // Setup all queues
     await Promise.all(Object.values(MESSAGE_QUEUES).map(qname => queue.createQueue(qname).catch(e => console.error(e))));
-  },
+  }
 
   async setupForwarding(sourceQueueName, targetQueues) {
     // SETUP LOGIC
     queue.subscribe(sourceQueueName, async message => {
       await Promise.all(targetQueues.map(targetQueueName => queue.sendMessage(targetQueueName, message.message)));
-      await queue.deleteMessage(MESSAGE_QUEUES.PARENT, message.id);
+      await queue.deleteMessage(targetQueues, message.id);
     });
-  },
-};
+  }
+}
+
+const broker = new MessageBroker();
 
 module.exports = broker;
